refactoring dropdown components
